+++ components/fields/index.ts
@@ -1,6 +1,6 @@
-export * from './CheckboxInput';
-export * from './FormField';
-export * from './FunctionInput';
-export * from './NumberInput';
-export * from './Select';
-export * from './TextInput';
+export * from './checkbox-input';
+export * from './form-field';
+export * from './function-input';
+export * from './number-input';
+export * from './select';
+export * from './text-input';
