+++ components/forms/index.ts
@@ -1,8 +1,8 @@
-export * from './AdvertismentForm';
-export * from './AdvertismentListForm';
-export * from './ControlBarForm';
-export * from './LayoutForm';
-export * from './LogoForm';
-export * from './MiniPlayerForm';
-export * from './StaticPreviewForm';
-export * from './TimelinePreviewForm';
+export * from './advertisment-form';
+export * from './advertisment-list-form';
+export * from './control-bar-form';
+export * from './layout-form';
+export * from './logo-form';
+export * from './mini-player-form';
+export * from './static-preview-form';
+export * from './timeline-preview-form';
