+++ components/player/index.ts
@@ -1,1 +1,1 @@
-export * from './AddEdit';
+export * from './add-edit';
